+++ moabb/datasets/bbci_eeg_fnirs.py
@@ -192,15 +192,6 @@ class BaseShin2017(BaseDataset):
 class Shin2017A(BaseShin2017):
     """Motor Imagey Dataset from Shin et al 2017.
 
-    .. admonition:: Dataset summary
-
-
-        =========  =======  =======  ==========  =================  ============  ===============  ===========
-        Name         #Subj    #Chan    #Classes    #Trials / class  Trials len    Sampling rate      #Sessions
-        =========  =======  =======  ==========  =================  ============  ===============  ===========
-        Shin2017A       29       30           2                 30  10s           200Hz                      3
-        =========  =======  =======  ==========  =================  ============  ===============  ===========
-
     Dataset from [1]_.
 
 
@@ -314,15 +305,6 @@ class Shin2017A(BaseShin2017):
 
 class Shin2017B(BaseShin2017):
     """Mental Arithmetic Dataset from Shin et al 2017.
-
-    .. admonition:: Dataset summary
-
-
-        =========  =======  =======  ==========  =================  ============  ===============  ===========
-        Name         #Subj    #Chan    #Classes    #Trials / class  Trials len    Sampling rate      #Sessions
-        =========  =======  =======  ==========  =================  ============  ===============  ===========
-        Shin2017B       29       30           2                 30  10s           200Hz                      3
-        =========  =======  =======  ==========  =================  ============  ===============  ===========
 
     Dataset from [1]_.