+++ docs/source/whats_new.rst
@@ -18,10 +18,13 @@ Develop branch
 Enhancements
 ~~~~~~~~~~~~
 - Add possibility to use OptunaGridSearch (:gh:`630` by `Igor Carrara`_)
+- Centralize dataset summary tables in CSV files (:gh:`635` by `Pierre Guetschel`_)
+- Add new dataset :class:`moabb.datasets.Liu2024` dataset (:gh:`619` by `Taha Habib`_)
+
 
 Bugs
 ~~~~
-- None
+- Fix caching in the workflows (:gh:`632` by `Pierre Guetschel`_)
 
 API changes
 ~~~~~~~~~~~
@@ -473,3 +476,4 @@ API changes
 .. _Brian Irvine: https://github.com/brianjohannes
 .. _Bruna Lopes: https://github.com/brunaafl
 .. _Yash Chauhan: https://github.com/jiggychauhi
+.. _Taha Habib: https://github.com/tahatt13